% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcalg.R
\name{ipc}
\alias{ipc}
\title{Estimate an IDAG}
\usage{
ipc(pc.fit, data, x, y, alpha = 0.01)
}
\arguments{
\item{pc.fit}{fitted PC object}

\item{data}{dataset}

\item{x}{intervention}

\item{y}{outcome}

\item{alpha}{significance level for regressions}
}
\value{
fitted IPC object
}
\description{
Estimate an IDAG for the causal effect of x on y from a fitted PC function.
}
\examples{
library(tibble)
library(pcalg)
n <- 1000
data <- tibble::tibble(
  A = rnorm(n),
  B = rnorm(n),
  Y = A + B + A * B + rnorm(n)
)

pc.fit <- pc(
  suffStat = list(C = cor(data), n = nrow(data)),
  indepTest = gaussCItest,
  alpha = 0.01,
  labels = colnames(data)
)

ipc.fit <- ipc(pc.fit, data, x = 'A', y = 'Y', alpha = 0.01)
}
